+++ CHANGELOG.md
@@ -5,7 +5,11 @@
   - [#98](https://github.com/Datatamer/unify-client-python/issues/98) Add `__geo_interface__` to Dataset
   - [#100](https://github.com/Datatamer/unify-client-python/issues/100) Add `from_geo_features` to Dataset
   - [#116](https://github.com/Datatamer/unify-client-python/issues/116) Add support for associating a dataset with a project
-
+  - [#109](https://github.com/Datatamer/unify-client-python/issues/109) Add support for profiling datasets
+  - [#86](https://github.com/Datatamer/unify-client-python/issues/86) Add support for creating projects
+  - [#114](https://github.com/Datatamer/unify-client-python/issues/114) Add support for generating pairs estimate
+  - [#106](https://github.com/Datatamer/unify-client-python/issues/106) Add support for initializing a source dataset
+  
   **BUG FIXES**
   - [#118](https://github.com/Datatamer/unify-client-python/issues/118) Fix JSON sent for Dataset.update_records